LayoutTests/editing/editability/set-inherit-parent-true-expected.txt

When contentEditable is set to "inherit", the attribute must be removed and element.contentEditable should return "inherit".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".





PASS document.getElementById("p").hasAttribute("contentEditable") is false
FAIL document.getElementById("p").contentEditable should be inherit. Was false.
FAIL document.getElementById("p").isContentEditable should be true. Was false.
FIXME: setContentEditable with true/false/inherit string is not working properly. https://bugs.webkit.org/show_bug.cgi?id=52058
PASS window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify") is "read-write"

PASS successfullyParsed is true

TEST COMPLETE
